% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{calES}
\alias{calES}
\title{Calculate expected scores}
\usage{
calES(ipar, theta = seq(-4, 4, 0.1))
}
\arguments{
\item{ipar}{a data frame containing estimated GRM parameters}

\item{theta}{a numeric indicating the range of theta}
}
\description{
Calculate expected scores
}
