% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{calProb}
\alias{calProb}
\title{Calculate probability}
\usage{
calProb(ipar, theta = seq(-4, 4, 0.1))
}
\arguments{
\item{ipar}{a data frame containing estimated GRM parameters}

\item{theta}{a numeric indicating the range of theta}
}
\description{
Calculate probability
}
