% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{ESplot}
\alias{ESplot}
\title{Plot expected scores by items}
\usage{
ESplot(
  fit,
  selected_item,
  theta,
  base_size = 16,
  line_size = 1,
  cal_option = "D"
)
}
\description{
Plot expected scores by items
}
