% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_out_tidy.r
\name{extract_est}
\alias{extract_est}
\title{Clean output to look like Mplus}
\usage{
extract_est(fit)
}
\arguments{
\item{fit}{an object from \code{\link{runGRM}}}
}
\description{
Clean output to look like Mplus
}
\details{
This tidies the results in the CFA framework.
}
