% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_out_tidy.r
\name{extract_fit}
\alias{extract_fit}
\title{Extract model fit}
\usage{
extract_fit(fit)
}
\arguments{
\item{fit}{an object from \code{\link{runGRM}}}
}
\value{
a list of IRT fit information
}
\description{
Extract model fit
}
\details{
This extracts model fit.
}
