% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{FSplot}
\alias{FSplot}
\title{Plot Factor score}
\usage{
FSplot(
  fit,
  type = "histogram",
  hist_bins = 20,
  fill_colour = "grey70",
  base_size = 16
)
}
\description{
Plot Factor score
}
\examples{
FSplot(fs_scores, type = "histogram", hist_bins = 20, base_size = 16)
FSplot(fs_scores, type = "density", hist_bins = 20, base_size = 16)
}
