% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.r
\name{genData}
\alias{genData}
\title{Generate IRT data}
\usage{
genData(eta, ipar)
}
\arguments{
\item{eta}{a matrix indicating individual true latent traits}

\item{ipar}{a dataframe containing item parameters}
}
\value{
a data frame containing graded response model data
}
\description{
Generate IRT data
}
\examples{

ipar <- genIRTpar(20, ncat = 3, 2)
eta <- genTheta(50, 2)
genData(eta, ipar)

}
