% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.r
\name{genTheta}
\alias{genTheta}
\title{Generate individual true latent traits}
\usage{
genTheta(nsample, nfac, l.cov = NULL)
}
\arguments{
\item{nsample}{a numeric indicating the number of people}

\item{nfac}{a numeric indicating the number of factors}

\item{l.cov}{a matrix containing latent covariances}
}
\value{
a matrix containing individual latent scores
}
\description{
Generate individual true latent traits
}
\details{
This generates individual latent scores
}
