% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{ICCplot}
\alias{ICCplot}
\title{Plot ICC or OCC}
\usage{
ICCplot(
  fit,
  selected_item,
  theta = seq(-3, 3, 0.1),
  plot.ps = FALSE,
  addlabel = F,
  base_size = 16,
  line_size = 1,
  cal_option = "D"
)
}
\description{
Plot ICC or OCC
}
\examples{
ICCplot(ipar, 1, seq(-3, 3, .1), plot.ps = FALSE, base_size = 16)
ICCplot(ipar, selected_item=2:4, theta=seq(-3, 3, .1), plot.ps = T, base_size = 16)
}
