% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{infoPlot}
\alias{infoPlot}
\title{Calculate item information}
\usage{
infoPlot(
  fit,
  selected_item,
  type = "icc",
  theta = seq(-3, 3, 0.1),
  base_size = 16,
  line_size = 1,
  cal_option = "D"
)
}
\description{
Calculate item information
}
\examples{
infoPlot(ipar, selected_item=1:4, type = "icc", base_size = 16)
infoPlot(ipar, selected_item=1:2, type = "tcc", base_size = 16)
}
