% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_out_tidy.r
\name{lavaan_output_cleaning}
\alias{lavaan_output_cleaning}
\title{clean lavaan class}
\usage{
lavaan_output_cleaning(lav.fit)
}
\arguments{
\item{lav.fit}{an obejct from \code{\link{lavaan}} package}
}
\description{
clean lavaan class
}
