% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_shinyapp.r
\name{shiny_server}
\alias{shiny_server}
\title{The Shiny App Server}
\usage{
shiny_server(input, output, session)
}
\arguments{
\item{input}{a shiny input}

\item{output}{a shiny output}

\item{session}{a seesion}
}
\description{
The Shiny App Server
}
