% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_grm.r
\name{trans_to_grm}
\alias{trans_to_grm}
\title{Transform SEM params into GRM params}
\usage{
trans_to_grm(grm.fit)
}
\arguments{
\item{grm.fit}{a matrix containing the estimates of GRM}
}
\description{
Transform SEM params into GRM params
}
