% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{ESplot}
\alias{ESplot}
\title{Plot expected scores by items}
\usage{
ESplot(
  fit,
  selected_item,
  theta = seq(-4, 4, 0.1),
  base_size = 16,
  line_size = 1,
  cal_option = "D"
)
}
\arguments{
\item{fit}{an object from \code{\link{runGRM}}}

\item{selected_item}{a numeric indicating for what items the function makes plots}

\item{theta}{a numeric indicating latent traints}

\item{base_size}{a numeric indicating the base font size}

\item{line_size}{a numeric indicating the size of line}

\item{cal_option}{a character indicating the plot color specified in
\code{\link{scale_color_viridis_d}} (default = \code{D})}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Plot expected scores by items
}
