% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.r
\name{genData}
\alias{genData}
\title{Generate IRT data}
\usage{
genData(eta, ipar)
}
\description{
Generate IRT data
}
\examples{

ipar <- genIRTpar(20, ncat = 3, 2)
eta <- MASS::mvrnorm(100, rep(0, 2), matrix(c(1,0,0,1),ncol=2))
genData(eta, ipar)

}
