% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.r
\name{genIRTpar}
\alias{genIRTpar}
\title{Generate Item parameters}
\usage{
genIRTpar(nitem = 25, ncat = 4, nfac = 3)
}
\arguments{
\item{nitem}{a numeric indicating the number of items}

\item{ncat}{a numeric indicating the number of categories}

\item{nfac}{a numeric indicating the number of factors}
}
\description{
Generate Item parameters
}
\examples{
genIRTpar(20, 4, 3)
}
