% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_grm.r
\name{genLavSyn}
\alias{genLavSyn}
\title{Generate lavaan syntax}
\usage{
genLavSyn(dat, nfac = 1)
}
\arguments{
\item{dat}{a data frame containing graded response data}

\item{nfac}{a numeric indicating the number of factors}
}
\description{
Generate lavaan syntax
}
