% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_grm.r
\name{runGRM}
\alias{runGRM}
\title{Run graded response model}
\usage{
runGRM(dat, lav.syntax, estimator)
}
\arguments{
\item{dat}{a data frame containing graded response model}

\item{lav.syntax}{a character indicating lavvan syntax}

\item{estimator}{a character indicating the type of estimator}
}
\description{
Run graded response model
}
