% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_runAPP.r
\name{startGRshiny}
\alias{startGRshiny}
\title{Start GRShiny}
\usage{
startGRshiny()
}
\value{
It launches the Shiny app for Graded Response Model.
}
\description{
An interactive Shiny application for running a GRM analysis.
}
\details{
\code{\link{startGRshiny}} is a caller function to open the Shiny interface of GRM.

This starts the IRT Shiny application on the user's local computer.
}
\examples{
if (interactive()) {
  startGRshiny()
}
}
\keyword{IRT}
