% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_runAPP.r
\name{startGRshiny}
\alias{startGRshiny}
\title{This function will start GRShiny}
\usage{
startGRshiny()
}
\description{
An interactive Shiny application for running a GRM analysis.
}
\details{
Start GRShiny

This starts the IRT Shiny application on the users local computer.
}
\examples{
\dontrun{
library(shiny)
startGRshiny()
}
}
\keyword{IRT}
